Added scalability, removed stupidity
Added with/as back in since the files are automatically closed that way. 
allitems.csv currently only uses first column, these values are added to a dictionary that is eventually written to the g_info file. Editing this will update each g_info when the account is logged into. 
table.csv is a placeholder loot table, the probability and type of fish caught is determined by this file. 
Made some changes to the way login info is input and displayed.