Updated equipment functions
Changed login effects on equipment to be a nested object instead of nested array/object yuckiness
Added boost to f_j generated by equipped items

Need to add modularity to the MultDict in the relog function so that any item can be put in. Probably select items from a list that regenerate over time.